+++ gantry/constants.py
@@ -17,3 +17,5 @@ RESULTS_DIR = "/results"
 METRICS_FILE = f"{RESULTS_DIR}/metrics.json"
 
 NFS_MOUNT = "/net/nfs.cirrascale"
+
+CLUSTERS_WITHOUT_NFS = {"ai2/jupiter-cirrascale"}