+++ gantry/util.py
@@ -406,7 +406,16 @@ def build_experiment_spec(
     if install is not None:
         task_spec = task_spec.with_env_var(name="INSTALL_CMD", value=install)
 
-    if nfs is None and clusters and all(["cirrascale" in cluster for cluster in clusters]):
+    if (
+        nfs is None
+        and clusters
+        and all(
+            [
+                "cirrascale" in cluster and cluster not in constants.CLUSTERS_WITHOUT_NFS
+                for cluster in clusters
+            ]
+        )
+    ):
         nfs = True
 
     if nfs: