+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Don't automatically attach NFS to "jupiter" cluster.
+
 ## [v0.23.1](https://github.com/allenai/beaker-gantry/releases/tag/v0.23.1) - 2024-05-14
 
 ### Changed